requirements.txt
fastapi
uvicorn[standard]
pydantic
numpy
faiss-cpu
sentence-transformers
python-dotenv
requests
django
django
requests
python-dotenv